requirements.txt
fastapi
uvicorn
langchain
streamlit
faiss-cpu
tiktoken
requests
langchain-community
python-dotenv
langchain-mistralai